use async/await for database models

+++ server/src/routes/acp/acpuser/ACPDeleteACPUserRoute.ts
@@ -19,6 +19,9 @@ class ACPDeleteACPUserRoute extends Route {
     private async post(req: Request, res: Response): Promise<void> {
         try {
             let user = await ACPUser.get(req.body.name);
+            if (!user) {
+                throw new Error('ACP user not found');
+            }
             await user.delete();
             res.send({ success: true, message: `Successfully deleted ACP user with username '${req.body.name}'` });
         } catch (err) {
